feat(pb-ajax): added progress and button indicates state
Current version did not give any visual feedback once the long-running process was started. This adds a HTML progress element which will be shown in indeterministic state when button was clicked. The trigger button
will also be disabled and readonly to prevent further clicks.

Upon completion the progress will be hidden and the button returns to normal state.