fix(pb-facsimile): perform better with many pb-facs-links
If there are many (like 400) pb-facs-links, the old implementation happily sent something in the
order of 400*400 = 160.000 facsimiles to OSD. For every pb-facs-link that appears in the DOM all
facsimiles are 'opened' in OSD. Which in turn passed all those requests to the
endpoint serving manifests.

Debounce these to only do 400 requests: when all of the pb-facs-link elements are 'in'. Furthermore,
if they are all the same, send only one: the unique one.